Previus work: [llama.cpp#778](https://github.com/ggerganov/llama.cpp/pull/778)

Previously, the initiative to implement Flash Attention to improve inference performance in llama.cpp had already been introduced. However, it was assumed that this approach would yield the expected results on the CPU, for that reason, it was discarded, and no further follow-up was given.

Flash Attention is actually designed to enhance GPU resource utilization through the use of tensor cores and shared memory, which is 30 times faster than global memory (VRAM), reducing unnecessary readings and writings.

Implementing this algorithm is particularly challenging because it requires taking into account hardware limitations and conflicts that can degrade performance, even when everything is expected to be fine (a.k.a. shared memory banks conflicts).

Tasks to be carried out during the execution of this project:

- [x] Analyzing and understanding concisely how Flash Attention works. Even if it's just creating a C++ code that replicates the exact functionality without much complexity. From this base, we move on to parallelize and optimize.
- [ ] Optimize and support different models.
- [ ] Create the backward kernel for training.

### Run test:

```bash
# get the code
git clone https://github.com/Pints-App/llama.cpp.git
cd llama.cpp

# build
mkdir build
cd build
cmake .. -DLLAMA_CUBLAS=ON
cmake --build . --config Release

# Run test (get result)
bin/test-flash-attention

# Run test (compare backends)
bin/test-flash-attention comp
```