@JohannesGaessler 2024-01-18
>Create the backward kernel for training.

Currently I don't think llama.cpp has good support for training at all. I want to work on it at some point but implementing a backwards pass currently doesn't make sense I think.