@JohannesGaessler 2024-01-18
Since the target architectures are Ampere and Ada Lovelace consider that copying data from VRAM to SRAM is more efficient using [asynchronous data copies](https://docs.nvidia.com/cuda/cuda-c-programming-guide/index.html#asynchronous-data-copies). This is both faster and reduces register pressure. It also allows you to load data and do computations in parallel.  This optimization can be added later though. I will soon push a prototype to https://github.com/ggerganov/llama.cpp/pull/4801 which (I think) is close to optimal.