@JohannesGaessler 2024-01-24
Current performance level on my system: 

| GPU      | Model         | Test   |   t/s master |   t/s flash-attn-cuda |   Speedup |
|:---------|:--------------|:-------|-------------:|----------------------:|----------:|
| RTX 3090 | llama 7B Q8_0 | pp512  |      3871.86 |               2402.47 |      0.62 |
| RTX 3090 | llama 7B Q8_0 | tg128  |        85.61 |                 85.65 |      1.00 |